learn : Add auth middleware and ownership to our routes

+++ backend/03_blog_application/app.js
@@ -5,10 +5,14 @@ import crypto from "node:crypto";
 import { writeFile } from "node:fs/promises";
 import path from "path";
 import multer from "multer";
+import cookieParser from "cookie-parser";
+import authMiddleware from "./middleware/authMiddleware.js";
 
 const app = express(); // Object
 
 app.use(express.json());
+app.use(cookieParser())
+
 
 const storage = multer.diskStorage({
   destination: function (req, file, cb) {
@@ -40,19 +44,20 @@ app.get("/blogs/:id", (req, res) => {
 });
 
 // CREATE 
-app.post("/blogs", upload.single("image"), async (req, res) => {
+app.post("/blogs", authMiddleware,upload.single("image"), async (req, res) => {
   const { title, content, author } = req.body;
-  console.log(req.body);
+
   if (!title || !content || !author) {
     return res.status(400).json({ message: "All fields are requried" });
   }
 
   const newBlog = {
     id: crypto.randomUUID(),
+    userId : req.user.id,
     title,
     content,
     author,
-    like: 0,
+    likes: [],
     comment: [],
     image: req.file ? `/uploads/` : null,
     createdAt: new Date().toISOString(),
@@ -70,9 +75,9 @@ app.post("/blogs", upload.single("image"), async (req, res) => {
 });
 
 // Blog UPdate -> PATCH
-app.patch("/blogs/:id", async (req, res) => {
-  const { id } = req.params; // id = 5b505468-0b7a-4501-b8d0-49f8a2ca7d3b
-  const blog = blogsData.find((blog) => blog.id === id);
+app.patch("/blogs/:id", authMiddleware ,async (req, res) => {
+  const { id } = req.params;
+  const blog = blogsData.find((blog) => blog.id === id &&  blog.userId === req.user.id);
 
   if (!blog) {
     return res.status(404).json({ message: "Blog not found" });
@@ -95,7 +100,7 @@ app.patch("/blogs/:id", async (req, res) => {
 });
 
 // Like
-app.post("/blogs/:id/likes", async (req, res) => {
+app.post("/blogs/:id/likes", authMiddleware,async (req, res) => {
   const { id } = req.params;
   const blog = blogsData.find((blog) => blog.id === id);
 
@@ -103,18 +108,28 @@ app.post("/blogs/:id/likes", async (req, res) => {
     return res.status(404).json({ message: "Blog not found" });
   }
 
-  blog.like += 1;
+  const alreadyLiked  = blog.likes.find((like) => {
+    return like.userId === req.user.id
+  })
+
+  if(alreadyLiked){
+    return res.json({message : "Already liked"})
+  }
+
+  blog.likes.push({
+    userId : req.user.id
+  })
 
   try {
     await writeFile("./blogsDB.json", JSON.stringify(blogsData, null, 2));
-    return res.status(201).json({ message: "Blog Liked" });
+    return res.status(201).json({ message: "Blog Liked" , blogCount : blog.likes.length });
   } catch (err) {
     return res.status(401).json({ message: "Failed to Like on a blog" });
   }
 });
 
 // Comment
-app.post("/blogs/:id/comment", async (req, res) => {
+app.post("/blogs/:id/comment", authMiddleware,async (req, res) => {
   const { id } = req.params;
   const blog = blogsData.find((blog) => blog.id === id);
 
@@ -122,13 +137,15 @@ app.post("/blogs/:id/comment", async (req, res) => {
     return res.status(404).json({ message: "Blog not found" });
   }
 
-  const { user, text } = req.body;
+  const { text } = req.body;
 
   const newComment = {
     id: crypto.randomUUID(),
-    user,
+    userId : req.user.id,
+    user: req.user.username,
     text,
     createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString()
   };
 
   blog.comment.push(newComment);
@@ -141,13 +158,14 @@ app.post("/blogs/:id/comment", async (req, res) => {
   }
 });
 
-// Blog Delete
-app.delete("/blogs/:id", async (req, res) => {
-  const { id } = req.params;
-  const blogIndex = blogsData.findIndex((blog) => blog.id === id); // 4
 
-  if (!blogIndex) {
-    return res.status(404).json({ message: "Blog not found" });
+// Blog Delete
+app.delete("/blogs/:id", authMiddleware,async (req, res) => {
+  const { id } = req.params;
+  const blogIndex = blogsData.findIndex((blog) => blog.id === id && blog.userId === req.user.id); // 4
+  
+  if (blogIndex=== -1) {
+    return res.status(404).json({ message: "Blog not found or unauthorized" });
   }
 
   blogsData.splice(blogIndex, 1);
@@ -193,14 +211,24 @@ app.post("/users/login" , (req , res) => {
     const {email , password} = req.body
 
     const user = usersData.find((user) => user.email === email && user.password === password)
+    
 
     if(!user){
       return res.status(404).json({message : "Invalid credentials"})
     }
 
+    res.cookie("uid", user.id)
+
     return res.status(200).json({message : "User logged in"})
 })
 
+
+// logout
+app.post("/users/logout" , (req , res) => {
+    res.clearCookie("uid")
+
+    return res.status(201).json({message : "Logged out"})
+})
 
 
 app.listen(7000, () => {
